Fix GetTDewPointFromVapPres in case of discontinuity with GetSatVapPres + bound HumRatio for all (#30)
Tightens up checks for handling the minimum value of humidity ratio (now reset to no lower than 1e-7) as well and it resolves the issue in GetTDewPointFromVapPres due to the discontinuity in the formulae in GetSatVapPres -- we restrict iteration to either left or right part of the saturation vapour pressure curve to avoid iterating back and forth across the discontinuity of the curve at the freezing point. When the partial pressure of water vapour is within the discontinuity of GetSatVapPres we simply return the freezing point of water.